sync worktree

+++ Modules/Basic/basic_strtoul.cpp
@@ -100,7 +100,7 @@ uint64_t ft_strtoul(const char *input_string, char **end_pointer, int32_t numeri
     if (end_pointer)
         *end_pointer = const_cast<char *>(current_character);
     if (sign_value < 0)
-        return (static_cast<uint64_t>(ULONG_MAX));
+        return (ULONG_MAX);
     if (overflow_detected)
         return (limit_value);
     return (accumulated_value);
